requirements.txt
Flask
Flask-SQLAlchemy
Flask-Script
Flask-Migrate
Flask-Assets
Flask-Mail
psycopg2-binary
schedule
daemon
lockfile
loguru
jsmin
cssmin
pika
pylint
